Task-1

+++ src/main/java/ru/netology/service/PostService.java
@@ -4,7 +4,9 @@ import ru.netology.exception.NotFoundException;
 import ru.netology.model.Post;
 import ru.netology.repository.PostRepository;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.concurrent.ConcurrentHashMap;
 
 public class PostService {
   private final PostRepository repository;
@@ -13,20 +15,25 @@ public class PostService {
     this.repository = repository;
   }
 
-  public List<Post> all() {
+  public ConcurrentHashMap<Long, Post> all() {
     return repository.all();
   }
 
   public Post getById(long id) {
-    return repository.getById(id).orElseThrow(NotFoundException::new);
+    return repository.getById(id);
   }
 
   public Post save(Post post) {
     return repository.save(post);
   }
 
-  public void removeById(long id) {
-    repository.removeById(id);
+  public Long removeById(long id) {
+    Long deletedId = getById(id).getId();
+    if (deletedId != null) {
+      repository.removeById(id);
+      return deletedId;
+    }
+    return -1L;
   }
 }
